requirements.txt
requests==2.31.0
azure-storage-blob==12.19.0
python-dotenv==1.0.0
PyYAML==6.0.1
urllib3==2.0.7 